Modify existentialAbstraction to avoid calls to contains.
The existentialAbstraction performs a transitive closure fix-point
algorithm, which at each iteration compares a list of "candidate"
parameters with a list of "already marked", which is a subset of it.
The test includes both equality (same object) and a call to the
"Type.contains" method, which creates a ContainsCollector object.

We modify the code, to move quick equality check before doing any
call to "Type.contains".