No check file for pos test
Optional check file for pos test is
brittle and was intentionally exploited
by only a couple of tests.

Also, it is confusing and error-prone
that an absent check file would have
different semantics for pos and neg.

Disallow check files for pos tests
and fail the test that has one,
requiring a neg test with `-Xfatal-warnings`
to verify warning text.